refactor: remove some graphql related modules
